Refactor client to sit behind interfaces
