+++ Keyspaces.java
@@ -1,0 +1,9 @@
+package grakn.client;
+
+import java.util.List;
+
+public interface Keyspaces {
+    void delete(String name);
+
+    List<String> retrieve();
+}
