+++ concept/Concept.java
@@ -19,7 +19,7 @@
 
 package grakn.client.concept;
 
-import grakn.client.GraknClient;
+import grakn.client.Transaction;
 import grakn.client.concept.impl.RuleImpl;
 import grakn.client.concept.thing.impl.AttributeImpl;
 import grakn.client.concept.thing.impl.EntityImpl;
@@ -212,7 +212,7 @@ public interface Concept<BaseType extends Concept<BaseType>> {
      * @param tx The transaction to use for the RPCs.
      * @return A remote concept using the given transaction to enable RPCs.
      */
-    Remote asRemote(GraknClient.Transaction tx);
+    Remote asRemote(Transaction tx);
 
     /**
      * Determine if the Concept is a SchemaConcept
@@ -383,7 +383,7 @@ public interface Concept<BaseType extends Concept<BaseType>> {
 
         @SuppressWarnings("unchecked")
         static <RemoteType extends Remote<BaseType>, BaseType extends Concept<BaseType>>
-        RemoteType of(ConceptProto.Concept concept, GraknClient.Transaction tx) {
+        RemoteType of(ConceptProto.Concept concept, Transaction tx) {
             ConceptId id = ConceptId.of(concept.getId());
             switch (concept.getBaseType()) {
                 case ENTITY:
