+++ concept/impl/SchemaConceptImpl.java
@@ -19,7 +19,7 @@
 
 package grakn.client.concept.impl;
 
-import grakn.client.GraknClient;
+import grakn.client.Transaction;
 import grakn.client.concept.Concept;
 import grakn.client.concept.ConceptId;
 import grakn.client.concept.Label;
@@ -56,7 +56,7 @@ public abstract class SchemaConceptImpl {
             extends ConceptImpl.Remote<BaseType>
             implements SchemaConcept.Remote<BaseType> {
 
-        public Remote(GraknClient.Transaction tx, ConceptId id) {
+        public Remote(Transaction tx, ConceptId id) {
             super(tx, id);
         }
