+++ concept/type/impl/TypeImpl.java
@@ -19,7 +19,7 @@
 
 package grakn.client.concept.type.impl;
 
-import grakn.client.GraknClient;
+import grakn.client.Transaction;
 import grakn.client.concept.Concept;
 import grakn.client.concept.ConceptId;
 import grakn.client.concept.Label;
@@ -62,7 +62,7 @@ public abstract class TypeImpl {
             extends SchemaConceptImpl.Remote<SomeRemoteType>
             implements Type.Remote<SomeRemoteType, SomeRemoteThing> {
 
-        protected Remote(GraknClient.Transaction tx, ConceptId id) {
+        protected Remote(Transaction tx, ConceptId id) {
             super(tx, id);
         }
