+++ test/behaviour/connection/session/SessionSteps.java
@@ -19,7 +19,7 @@
 
 package grakn.client.test.behaviour.connection.session;
 
-import grakn.client.GraknClient;
+import grakn.client.Session;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 
@@ -63,14 +63,14 @@ public class SessionSteps {
 
     @Then("session(s) is/are null: {bool}")
     public void sessions_are_null(Boolean isNull) {
-        for (GraknClient.Session session : sessions) {
+        for (Session session : sessions) {
             assertEquals(isNull, isNull(session));
         }
     }
 
     @Then("session(s) is/are open: {bool}")
     public void sessions_are_open(Boolean isOpen) {
-        for (GraknClient.Session session : sessions) {
+        for (Session session : sessions) {
             assertEquals(isOpen, session.isOpen());
         }
     }
@@ -99,7 +99,7 @@ public class SessionSteps {
     @Then("session(s) has/have keyspace(s):")
     public void sessions_have_keyspaces(List<String> names) {
         assertEquals(names.size(), sessions.size());
-        Iterator<GraknClient.Session> sessionIter = sessions.iterator();
+        Iterator<Session> sessionIter = sessions.iterator();
 
         for (String name : names) {
             assertEquals(name, sessionIter.next().keyspace().name());
@@ -109,7 +109,7 @@ public class SessionSteps {
     @Then("sessions in parallel have keyspaces:")
     public void sessions_in_parallel_have_keyspaces(List<String> names) {
         assertEquals(names.size(), sessionsParallel.size());
-        Iterator<CompletableFuture<GraknClient.Session>> futureSessionIter = sessionsParallel.iterator();
+        Iterator<CompletableFuture<Session>> futureSessionIter = sessionsParallel.iterator();
         CompletableFuture[] assertions = new CompletableFuture[names.size()];
 
         int i = 0;
