+++ test/deployment/src/test/java/application/MavenApplicationTest.java
@@ -1,6 +1,8 @@
 package application;
 
 import grakn.client.GraknClient;
+import grakn.client.Session;
+import grakn.client.Transaction;
 import grakn.client.answer.ConceptMap;
 import graql.lang.Graql;
 import org.junit.Test;
@@ -13,9 +15,9 @@ import static org.junit.Assert.assertEquals;
 public class MavenApplicationTest {
     @Test
     public void testImport() {
-        GraknClient client = new GraknClient("localhost:48555");
-        GraknClient.Session session = client.session("grakn");
-        GraknClient.Transaction tx = session.transaction().write();
+        GraknClient client = GraknClient.open("localhost:48555");
+        Session session = client.session("grakn");
+        Transaction tx = session.transaction().write();
         List<ConceptMap> answers = tx.execute(Graql.match(var("t").sub("thing")).get());
         tx.close();
         session.close();
